refactor(ch-lv2): 클래스들을 application, domain, view 3개의 패키지로 분리
