+++ src/main/java/io/github/seonrizee/kiosk/challenge/lv2/Main.java
@@ -1,5 +1,8 @@
 package io.github.seonrizee.kiosk.challenge.lv2;
 
+import io.github.seonrizee.kiosk.challenge.lv2.application.Kiosk;
+import io.github.seonrizee.kiosk.challenge.lv2.domain.Menu;
+import io.github.seonrizee.kiosk.challenge.lv2.domain.MenuItem;
 import java.util.List;
 
 public class Main {
