+++ src/main/java/io/github/seonrizee/kiosk/challenge/lv2/domain/Cart.java
@@ -1,4 +1,4 @@
-package io.github.seonrizee.kiosk.challenge.lv2;
+package io.github.seonrizee.kiosk.challenge.lv2.domain;
 
 import java.util.LinkedHashMap;
 import java.util.List;
