+++ src/main/java/io/github/seonrizee/kiosk/challenge/lv2/domain/Menu.java
@@ -1,4 +1,4 @@
-package io.github.seonrizee.kiosk.challenge.lv2;
+package io.github.seonrizee.kiosk.challenge.lv2.domain;
 
 import java.util.List;
