+++ src/main/java/io/github/seonrizee/kiosk/challenge/lv2/domain/MenuItem.java
@@ -1,4 +1,4 @@
-package io.github.seonrizee.kiosk.challenge.lv2;
+package io.github.seonrizee.kiosk.challenge.lv2.domain;
 
 public class MenuItem {
